feat: move selection logic into Card and remove CharacterSelectItemCheckbox component. Update app-data constants accordingly
